+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Change Log
 
+### v0.9.1
+
+- More documentation
+- GitHub Actions CI
+- Support for parsing "OFFSET # ROWS"
+- Support for parsing "FETCH NEXT # ROWS ONLY"
+
 ### v0.9.0
 
 - Initial release with `SelectBuilder`